LPD-55736 Move to specific util
